+++ src/sdc11073/mdib/mdibbase.py
@@ -358,33 +358,34 @@ class MdibBase:
                                              attrib={'DescriptionVersion': str(self.mddescription_version)},
                                              nsmap=doc_nsmap)
         for root_container in root_containers:
-            node = self.make_descriptor_node(root_container, tag=pm.Mds, set_xsi_type=False)
-            md_description_node.append(node)
+            self.make_descriptor_node(root_container, md_description_node, tag=pm.Mds, set_xsi_type=False)
         return md_description_node
 
     def make_descriptor_node(self,
                              descriptor_container: AbstractDescriptorContainer,
+                             parent_node: xml_utils.LxmlElement,
                              tag: etree_.QName,
                              set_xsi_type: bool = True) -> xml_utils.LxmlElement:
         """Create a lxml etree node with subtree from instance data.
 
         :param descriptor_container: a descriptor container instance
+        :param parent_node: parent node
         :param tag: tag of node
         :param set_xsi_type: if true, the NODETYPE will be used to set the xsi:type attribute of the node
         :return: an etree node.
         """
         ns_map = self.nsmapper.partial_map(self.nsmapper.PM, self.nsmapper.XSI) \
             if set_xsi_type else self.nsmapper.partial_map(self.nsmapper.PM)
-        node = etree_.Element(tag,
-                              attrib={'Handle': descriptor_container.Handle},
-                              nsmap=ns_map)
+        node = etree_.SubElement(parent_node,
+                                 tag,
+                                 attrib={'Handle': descriptor_container.Handle},
+                                 nsmap=ns_map)
         descriptor_container.update_node(node, self.nsmapper, set_xsi_type)  # create all
         child_list = self.descriptions.parent_handle.get(descriptor_container.Handle, [])
         # append all child containers, then bring all child elements in correct order
         for child in child_list:
             child_tag, set_xsi = descriptor_container.tag_name_for_child_descriptor(child.NODETYPE)
-            child_node = self.make_descriptor_node(child, child_tag, set_xsi)
-            node.append(child_node)
+            self.make_descriptor_node(child, node, child_tag, set_xsi)
         descriptor_container.sort_child_nodes(node)
         return node